Ensure rakefile will load if ORACLE_HOME isn't set.

+++ Rakefile
@@ -201,21 +201,25 @@ namespace :deploy do
   end
 end
 
-Bcdatabase::Oracle.create_users_task(["cc_pers_test"])
-desc "Import the cc_pers_test dump (requires imp)"
-Bcdatabase::Oracle.import_task(
-  "test:db:import",
-  :local_oracle,
-  :local_oracle,
-  :cc_pers_test,
-  :filename => "db/exports/cc_pers_test.dmp"
-)
-Bcdatabase::Oracle.wipe_task(
-  "test:db:wipe",
-  :local_oracle,
-  :cc_pers_test
-)
-task "test:db:import" => "test:db:wipe"
+if ENV['ORACLE_HOME']
+  Bcdatabase::Oracle.create_users_task(["cc_pers_test"])
+  desc "Import the cc_pers_test dump (requires imp)"
+  Bcdatabase::Oracle.import_task(
+    "test:db:import",
+    :local_oracle,
+    :local_oracle,
+    :cc_pers_test,
+    :filename => "db/exports/cc_pers_test.dmp"
+    )
+  Bcdatabase::Oracle.wipe_task(
+    "test:db:wipe",
+    :local_oracle,
+    :cc_pers_test
+    )
+  task "test:db:import" => "test:db:wipe"
+else
+  $stderr.puts "Not defining bcoracle tasks because ORACLE_HOME is not set."
+end
 
 namespace :ci do
   task :all => [:spec, :cucumber]
